requirements.txt
# Core dependencies
streamlit>=1.40.0
python-dotenv>=1.0.0

# AI APIs
anthropic>=0.40.0
google-genai>=1.0.0
google-generativeai>=0.8.0
openai>=1.30.0

# Repository analysis
gitpython>=3.1.0
requests>=2.32.0

# Data handling
pyyaml>=6.0.0
# sqlite3 is built-in to Python

# Google Cloud (for Vertex AI)
google-cloud-aiplatform>=1.50.0
google-auth>=2.29.0

# Utilities
# pathlib is built-in to Python
